moving each page to its own folder for future involvement with individual scss files
