+++ client/src/pages/SingleCountryCont/SingleCountryCont.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import Header from '../components/Header';
-import CountryCards from '../components/CountryCards';
-import Footer from '../components/Footer';
-import SingleCountryHeader from '../components/SingleCountryHeader'
-import SingleCountry from '../components/SingleCountry';
+import Header from '../../components/Header';
+import CountryCards from '../../components/CountryCards';
+import Footer from '../../components/Footer';
+import SingleCountryHeader from '../../components/SingleCountryHeader'
+import SingleCountry from '../../components/SingleCountry';
