+++ client/src/pages/Splash/Splash.scss
@@ -1,5 +1,5 @@
 .body{
-    background-image: url("../../public/images/homeBackground.png");
+    background-image: url("../../../public/images/homeBackground.png");
     background-size: cover;
     // background-color: #04566e;
     background-repeat: no-repeat;
